+++ packages/smarthr-ui/src/components/ComboBox/MultiComboBox/index.tsx
@@ -1,0 +1,1 @@
+export { MultiComboBox } from './MultiComboBox'
